Prevent a circular import
