+++ bellows/ezsp/config.py
@@ -1,0 +1,76 @@
+from __future__ import annotations
+
+import dataclasses
+
+import bellows.ezsp.v4.types as types_v4
+import bellows.ezsp.v6.types as types_v6
+import bellows.types as t
+
+
+@dataclasses.dataclass(frozen=True)
+class RuntimeConfig:
+    config_id: t.enum8
+    value: int
+    minimum: bool = False
+
+
+DEFAULT_CONFIG = [
+    RuntimeConfig(
+        config_id=types_v4.EzspConfigId.CONFIG_SOURCE_ROUTE_TABLE_SIZE,
+        value=200,
+        minimum=True,
+    ),
+    RuntimeConfig(
+        config_id=types_v4.EzspConfigId.CONFIG_END_DEVICE_POLL_TIMEOUT,
+        value=8,
+    ),
+    RuntimeConfig(
+        config_id=types_v4.EzspConfigId.CONFIG_INDIRECT_TRANSMISSION_TIMEOUT,
+        value=7680,
+    ),
+    RuntimeConfig(
+        config_id=types_v4.EzspConfigId.CONFIG_STACK_PROFILE,
+        value=2,
+    ),
+    RuntimeConfig(
+        config_id=types_v4.EzspConfigId.CONFIG_SUPPORTED_NETWORKS,
+        value=1,
+        minimum=True,
+    ),
+    RuntimeConfig(
+        config_id=types_v4.EzspConfigId.CONFIG_MULTICAST_TABLE_SIZE,
+        value=16,
+        minimum=True,
+    ),
+    RuntimeConfig(
+        config_id=types_v4.EzspConfigId.CONFIG_TRUST_CENTER_ADDRESS_CACHE_SIZE,
+        value=2,
+        minimum=True,
+    ),
+    RuntimeConfig(
+        config_id=types_v4.EzspConfigId.CONFIG_SECURITY_LEVEL,
+        value=5,
+    ),
+    RuntimeConfig(
+        config_id=types_v4.EzspConfigId.CONFIG_ADDRESS_TABLE_SIZE,
+        value=16,
+        minimum=True,
+    ),
+    RuntimeConfig(
+        config_id=types_v6.EzspConfigId.CONFIG_TC_REJOINS_USING_WELL_KNOWN_KEY_TIMEOUT_S,
+        value=90,
+    ),
+    RuntimeConfig(
+        config_id=types_v4.EzspConfigId.CONFIG_PAN_ID_CONFLICT_REPORT_THRESHOLD,
+        value=2,
+    ),
+    RuntimeConfig(
+        config_id=types_v4.EzspConfigId.CONFIG_APPLICATION_ZDO_FLAGS,
+        value=(
+            t.EmberZdoConfigurationFlags.APP_RECEIVES_SUPPORTED_ZDO_REQUESTS
+            | t.EmberZdoConfigurationFlags.APP_HANDLES_UNSUPPORTED_ZDO_REQUESTS
+        ),
+    ),
+    # Must be set last
+    RuntimeConfig(types_v4.EzspConfigId.CONFIG_PACKET_BUFFER_COUNT, value=0xFF),
+]
